requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78  # This version is available
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.13.0
scikit-learn==1.3.0
albumentations==1.3.1
plotly==5.15.0
pymongo==4.5.0
python-multipart==0.0.6
kaggle==1.5.16
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1